pyproject.toml
[tool.poetry]
name = "the-excel-servant"
version = "0.1.0"
description = ""
authors = ["yuxt <yuxt@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.11"
pandas = "1.5.3"
pyarrow = "13.0.0"
xlwings = "0.30.12"
streamlit = "1.28.2"

[tool.poetry.group.dev.dependencies]
pytest-dotenv = "0.5.2"
pytest = "7.4.3"
pre-commit = "3.5.0"
ipykernel = "6.27.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
